[#4] MusicKit 관련 파일들 기본 세팅
