+++ AGAMI/Sources/Service/MusicService.swift
@@ -1,0 +1,8 @@
+//
+//  MusicService.swift
+//  AGAMI
+//
+//  Created by Seoyeon Choi on 10/11/24.
+//
+
+import Foundation
